Update entity.js
This bug is a very old one. It even exists in nepphhh/arrasio, so it is a mistake by neph.
Players have noticed that sometimes, tanks and bullets dashes forward when they get hit. Let's test with modified eggs. Have their health buffed to 65535, regen nerfed to 0, density buffed to 65535, pushability set to 0 if it isn't already 0, ai('moveincircles') removed, and hits_own_type changed to 'never'. Buff the reload of generators by 10x. Then, set up the server and spawn a line of eggs, not too much or they will(unlikely) cause lags. Then, switch to a different team, and ram them, or shoot a bullet into them. Your tank/bullet will move notably faster, but only if they are still colliding with the eggs. That's because stepRemaining don't work properly. It is supposed to record the amount of early movements done to entity by advancedCollision(), and do the remaining movements later when the variable is reset. However, the "remaining movements" are done AFTER the stepRemaining variable is reset. Thus, they always do full movement while the advancedCollision() function still pushes the moving entity to where it would collide to other entities with a limit of 1.